\docType{data}
\name{barley}
\alias{barley}
\title{Barley Gene Expression Data}
\format{The data set "barley" is a matrix with 22841 rows and 75 columns.}
\source{
  \url{http://www.plexdb.org/modules/PD_browse/experiment_browser.php}.
}
\description{
  The data produced by experiment "Genetic regulation of
  gene expression of barley in response to stem rust (Pgt
  isolate TTKS) and can be access from PLEXdb (BB64). There
  is a file called ``BB64_RMA_tmt_medians.txt'' on the
  download page contains RMA expressions. The rma
  expression for the 75 chips involve fungas infection is
  the dataset barley in this package. Note that this data
  set is used for illustration in Bancroft et al. (2012).
}
\references{
  Tim Bancroft, Chuanlong Du and Dan Nettleton (2012)
  Estimation of False Discovery Rate Using Sequential
  Permutation P -Values.
}
\keyword{data}
\keyword{dataset}
\keyword{datasets}
\keyword{set}

