\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Acute Lymphoma Leukemia Data}
\format{The data set \code{leukemia} is a matrix with the first 5 columns being gene expression
for the 5 males who have a translocation between chromosome 4 and 22 (ALL1/AF4) and the last
21 columns being gene expression for the 21 males who have a translocation between chromosomes
9 and 22 (BCR/ABL).}
\description{
  The Band T-cell Acute Lymphocyctic Leukemia (ALL) data
  set can be access via the Bioconductor ALL package at
  \url{www.bioconductor.org}. Measures of messenger
  ribonucleic acid (mRNA)--commonly referred to as
  expression levels--are available for 12,625 probesets in
  128 ALL patients. Of these 128 patients, we focus on the
  21 males who have been classified as having a
  translocation between chromosomes 9 and 22 (BCR/ABL) and
  the 5 males who have a translocation between chromosomes
  4 and 11 (ALL1/AF4). This subset of data is the
  \code{leukemia} dataset in this package. Note that this
  dataset is used for illustrations in Bancroft et al.
  (2012).
}
\references{
  Tim Bancroft, Chuanlong Du and Dan Nettleton (2012)
  Estimation of False Discovery Rate Using Sequential
  Permutation P -Values.
}
\keyword{data}
\keyword{datasets}
\keyword{set}

