\name{parse_parameter}
\alias{parse_parameter}
\title{Parses the parameter list into C++ code.}
\usage{
  parse_parameter(par)
}
\arguments{
  \item{par}{a string vector with length 3. par[1]: name of
  the parameter par[2]: data structure of the parameter
  par[3]: type of data}
}
\description{
  Parses the parameter list into C++ code.
}

